added checksum tests when exporting a collection with replicate
The replicate install mode is meant to be used as a way to restructure
the mod archive's folder structure, checksum differences are not carried
over to the end user which is guaranteed to cause modding environment
mismatches between the user and the curator.

We no longer allow curators to export mods which have local changes,
using the replicate install mode.

closes nexus-mods/vortex#17368